[clean] clean old ChannelTransaction and ChannelAccount.
