+++ executor/src/block_processor.rs
@@ -804,7 +804,6 @@ where
                     match transaction.as_signed_user_txn()?.payload() {
                         TransactionPayload::Program
                         | TransactionPayload::Module(_)
-                        | TransactionPayload::Channel(_)
                         | TransactionPayload::ChannelV2(_)
                         | TransactionPayload::Script(_) => {
                             bail!("Write set should be a subset of read set.")
