+++ language/e2e-tests/src/account.rs
@@ -168,14 +168,6 @@ impl Account {
                 gas_unit_price,
                 Duration::from_secs(u64::max_value()),
             ),
-            TransactionPayload::Channel(channel_payload) => RawTransaction::new_channel(
-                *self.address(),
-                sequence_number,
-                channel_payload,
-                max_gas_amount,
-                gas_unit_price,
-                Duration::from_secs(u64::max_value()),
-            ),
             TransactionPayload::ChannelV2(channel_payload) => RawTransaction::new_channel_v2(
                 *self.address(),
                 sequence_number,
