+++ language/functional_tests/tests/testsuite/channel/channel_sequence_number.mvir
@@ -1,25 +1,38 @@
-//! account: bob
-//! account: alice
+//! account: alice, 10000000
+//! account: bob, 10000000
+//! channel: ch1, alice|bob
 
-module Test{
-    import 0x0.ChannelAccount;
-
-    public assert_channel_sequence_number(expect_sender_number: u64, expect_receiver_number:u64){
-        assert(ChannelAccount.sender_sequence_number() == move(expect_sender_number), 1024);
-        assert(ChannelAccount.receiver_sequence_number() == move(expect_receiver_number), 1025);
-        return;
-    }
+//! new-transaction
+//! sender: alice
+import 0x0.LibraAccount;
+main(){
+    LibraAccount.enable_channel();
+    return;
 }
 
 //! new-transaction
-//! sender: alice
-//! receiver: bob
+//! sender: bob
+import 0x0.LibraAccount;
+main(){
+    LibraAccount.enable_channel();
+    return;
+}
 
-0x0.ChannelAccount.open 100000 50000
 
 //! new-transaction
 //! sender: alice
-//! receiver: bob
+//! txn-channel: ch1, alice
+//! args: {{bob}},50000,40000
 
-{{default}}.Test.assert_channel_sequence_number 1 1
+0x0.ChannelScript.open
+
+//! new-transaction
+//! sender: alice
+//! txn-channel: ch1, alice
+import 0x0.LibraAccount;
+main(){
+    assert(LibraAccount.channel_sequence_number() == 1, 1024);
+    return;
+}
+
