+++ language/functional_tests/tests/testsuite/channel/channel_test.mvir
