+++ language/functional_tests/tests/testsuite/channel/gobang_test.mvir
@@ -428,7 +428,6 @@ main(){
 //! sender: alice
 //! txn-channel: ch1,alice
 //! args: {{bob}}
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main(receiver: address){
@@ -441,8 +440,6 @@ main(receiver: address){
 //! new-transaction
 //! sender: bob
 //! txn-channel: ch1,bob
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main(){
@@ -456,8 +453,6 @@ main(){
 //! sender: alice
 //! txn-channel: ch1,alice
 //! args: {{bob}},2,2
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main(receiver: address, x: u64, y: u64) {
@@ -473,8 +468,6 @@ main(receiver: address, x: u64, y: u64) {
 //! sender: bob
 //! txn-channel: ch1,bob
 //! args: {{alice}},3,2
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main(receiver: address, x: u64, y: u64) {
@@ -489,8 +482,6 @@ main(receiver: address, x: u64, y: u64) {
 //! sender: alice
 //! txn-channel: ch1,alice
 //! args: {{bob}},2,3
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main(receiver: address, x: u64, y: u64) {
@@ -505,8 +496,6 @@ main(receiver: address, x: u64, y: u64) {
 //! sender: bob
 //! txn-channel: ch1,bob
 //! args: {{alice}},3,3
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main(receiver: address, x: u64, y: u64) {
@@ -521,8 +510,6 @@ main(receiver: address, x: u64, y: u64) {
 //! sender: alice
 //! txn-channel: ch1,alice
 //! args: {{bob}},2,4
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main(receiver: address, x: u64, y: u64) {
@@ -537,8 +524,6 @@ main(receiver: address, x: u64, y: u64) {
 //! sender: bob
 //! txn-channel: ch1,alice
 //! args: {{alice}},3,4
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main(receiver: address, x: u64, y: u64) {
@@ -553,8 +538,6 @@ main(receiver: address, x: u64, y: u64) {
 //! sender: alice
 //! txn-channel: ch1,alice
 //! args: {{bob}},2,5
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main(receiver: address, x: u64, y: u64) {
@@ -569,8 +552,6 @@ main(receiver: address, x: u64, y: u64) {
 //! sender: bob
 //! txn-channel: ch1,alice
 //! args: {{alice}},3,5
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main(receiver: address, x: u64, y: u64) {
@@ -585,8 +566,6 @@ main(receiver: address, x: u64, y: u64) {
 //! sender: alice
 //! txn-channel: ch1,alice
 //! args: {{bob}},2,6
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main(receiver: address, x: u64, y: u64) {
@@ -602,8 +581,6 @@ main(receiver: address, x: u64, y: u64) {
 //! new-transaction
 //! sender: alice
 //! txn-channel: ch1,alice
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main() {
@@ -618,8 +595,6 @@ main() {
 //! new-transaction
 //! sender: bob
 //! txn-channel: ch1,bob
-
-import 0x0.ChannelAccount;
 import {{default}}.TheFive;
 
 main() {
