--- language/functional_tests/tests/testsuite/channel/is_channel_txn.mvir
@@ -1,38 +1,0 @@
-//! account: bob
-//! account: alice
-
-module Test{
-    import 0x0.ChannelAccount;
-    import 0x0.ChannelTransaction;
-
-    public assert_channel_txn(){
-        let receiver: address;
-        assert(ChannelTransaction.native_is_channel_txn() == true, 42);
-        receiver = ChannelTransaction.native_get_txn_receiver();
-        assert(copy(receiver) == {{bob}}, 42);
-        return;
-    }
-}
-
-//! new-transaction
-//! sender: alice
-import 0x0.ChannelTransaction;
-main() {
-    let x: bool;
-    x = ChannelTransaction.native_is_channel_txn();
-
-    assert(copy(x) == false, 77);
-    return;
-}
-
-//! new-transaction
-//! sender: alice
-//! receiver: bob
-
-0x0.ChannelAccount.open 100000 50000
-
-//! new-transaction
-//! sender: alice
-//! receiver: bob
-
-{{default}}.Test.assert_channel_txn
