+++ language/vm/src/transaction_metadata.rs
@@ -8,12 +8,6 @@ use libra_types::{
     transaction::{SignedTransaction, TransactionPayload},
 };
 
-pub struct ChannelMetadata {
-    pub receiver: AccountAddress,
-    pub channel_sequence_number: u64,
-    pub receiver_public_key: Ed25519PublicKey,
-}
-
 pub struct ChannelMetadataV2 {
     pub channel_address: AccountAddress,
     pub channel_sequence_number: u64,
@@ -30,21 +24,11 @@ pub struct TransactionMetadata {
     pub max_gas_amount: GasUnits<GasCarrier>,
     pub gas_unit_price: GasPrice<GasCarrier>,
     pub transaction_size: AbstractMemorySize<GasCarrier>,
-    pub channel_metadata: Option<ChannelMetadata>,
     pub channel_metadata_v2: Option<ChannelMetadataV2>,
 }
 
 impl TransactionMetadata {
     pub fn new(txn: &SignedTransaction) -> Self {
-        let channel_metadata = match txn.payload() {
-            TransactionPayload::Channel(channel_payload) => Some(ChannelMetadata {
-                receiver: channel_payload.receiver(),
-                channel_sequence_number: channel_payload.channel_sequence_number(),
-                receiver_public_key: channel_payload.receiver_public_key.clone(),
-            }),
-            _ => None,
-        };
-
         let channel_metadata_v2 = match txn.payload() {
             TransactionPayload::ChannelV2(channel_payload) => Some(ChannelMetadataV2 {
                 channel_address: channel_payload.channel_address(),
@@ -64,7 +48,6 @@ impl TransactionMetadata {
             max_gas_amount: GasUnits::new(txn.max_gas_amount()),
             gas_unit_price: GasPrice::new(txn.gas_unit_price()),
             transaction_size: AbstractMemorySize::new(txn.raw_txn_bytes_len() as u64),
-            channel_metadata,
             channel_metadata_v2,
         }
     }
@@ -93,22 +76,8 @@ impl TransactionMetadata {
         self.transaction_size
     }
 
-    pub fn receiver(&self) -> Option<AccountAddress> {
-        self.channel_metadata
-            .as_ref()
-            .map(|metadata| metadata.receiver)
-    }
-
-    pub fn channel_metadata(&self) -> Option<&ChannelMetadata> {
-        self.channel_metadata.as_ref()
-    }
-
     pub fn channel_metadata_v2(&self) -> Option<&ChannelMetadataV2> {
         self.channel_metadata_v2.as_ref()
-    }
-
-    pub fn is_channel_txn(&self) -> bool {
-        self.channel_metadata.is_some()
     }
 
     pub fn is_channel_txn_v2(&self) -> bool {
@@ -126,7 +95,6 @@ impl Default for TransactionMetadata {
             max_gas_amount: GasUnits::new(100_000_000),
             gas_unit_price: GasPrice::new(0),
             transaction_size: AbstractMemorySize::new(0),
-            channel_metadata: None,
             channel_metadata_v2: None,
         }
     }
