+++ language/vm/vm-runtime/vm-runtime-types/src/native_functions/dispatch.rs
@@ -292,100 +292,7 @@ lazy_static! {
             ],
             vec![]
         );
-        // Channel Account
-        add!(m, addr, "ChannelAccount", "native_move_to_channel",
-           |_, _| {
-                Err(VMStatus::new(StatusCode::UNREACHABLE).with_message(
-                            "native_move_to_channel does not have a native implementation"
-                                .to_string()))
-             },
-            vec![Kind::Resource],
-            vec![TypeParameter(0),
-                Bool
-            ],
-            vec![]
-        );
-        add!(m, addr, "ChannelAccount", "native_exist_channel",
-          |_, _| {
-                Err(VMStatus::new(StatusCode::UNREACHABLE).with_message(
-                            "native_exist_channel does not have a native implementation"
-                                .to_string()))
-             },
-            vec![Kind::Resource],
-            vec![Bool],
-            vec![Bool]
-        );
-        add!(m, addr, "ChannelAccount", "native_move_from_channel",
-          |_, _| {
-                Err(VMStatus::new(StatusCode::UNREACHABLE).with_message(
-                            "native_move_from_channel does not have a native implementation"
-                                .to_string()))
-             },
-            vec![Kind::Resource],
-            vec![Bool],
-            vec![TypeParameter(0)]
-        );
-        add!(m, addr, "ChannelAccount", "native_borrow_channel",
-            |_, _| {
-                Err(VMStatus::new(StatusCode::UNREACHABLE).with_message(
-                            "native_borrow_channel does not have a native implementation"
-                                .to_string()))
-             },
-            vec![Kind::Resource],
-            vec![Bool],
-            vec![MutableReference(Box::new(TypeParameter(0)))]
-        );
-        add!(m, addr, "ChannelTransaction", "native_is_offchain",
-            |_, _| {
-                Err(VMStatus::new(StatusCode::UNREACHABLE).with_message(
-                            "native_is_offchain does not have a native implementation"
-                                .to_string()))
-             },
-            vec![],
-            vec![],
-            vec![Bool]
-        );
-        add!(m, addr, "ChannelTransaction", "native_get_txn_receiver",
-            |_, _| {
-                Err(VMStatus::new(StatusCode::UNREACHABLE).with_message(
-                            "native_get_txn_receiver does not have a native implementation"
-                                .to_string()))
-             },
-            vec![],
-            vec![],
-            vec![Address]
-        );
-        add!(m, addr, "ChannelTransaction", "native_is_channel_txn",
-            |_, _| {
-                Err(VMStatus::new(StatusCode::UNREACHABLE).with_message(
-                            "native_is_offchain does not have a native implementation"
-                                .to_string()))
-             },
-            vec![],
-            vec![],
-            vec![Bool]
-        );
-        add!(m, addr, "ChannelTransaction", "native_get_txn_receiver_public_key",
-            |_, _| {
-                Err(VMStatus::new(StatusCode::UNREACHABLE).with_message(
-                            "native_get_txn_receiver_public_key does not have a native implementation"
-                                .to_string()))
-             },
-            vec![],
-            vec![],
-            vec![ByteArray]
-        );
-        add!(m, addr, "ChannelTransaction", "native_get_txn_channel_sequence_number",
-            |_, _| {
-                Err(VMStatus::new(StatusCode::UNREACHABLE).with_message(
-                            "native_get_txn_channel_sequence_number does not have a native implementation"
-                                .to_string()))
-             },
-            vec![],
-            vec![],
-            vec![U64]
-        );
-        //ChannelTransaction v2
+        //ChannelTransaction
         add!(m, addr, "ChannelTransaction", "is_offchain",
             |_, _| {
                 Err(VMStatus::new(StatusCode::UNREACHABLE).with_message(
