+++ types/src/transaction/helpers.rs
@@ -1,19 +1,17 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::transaction::ChannelTransactionPayloadBodyV2;
 use crate::{
     account_address::AccountAddress,
     proto::types::SignedTransaction as ProtoSignedTransaction,
-    transaction::{
-        ChannelScriptBody, ChannelWriteSetBody, RawTransaction, SignedTransaction,
-        TransactionPayload,
-    },
+    transaction::{RawTransaction, SignedTransaction, TransactionPayload},
 };
 use chrono::Utc;
 use failure::prelude::*;
 use libra_crypto::{
     ed25519::*,
-    hash::{CryptoHash, CryptoHasher, TestOnlyHash, TestOnlyHasher},
+    hash::{CryptoHash, TestOnlyHash},
     test_utils::KeyPair,
     traits::SigningKey,
     HashValue,
@@ -71,19 +69,9 @@ pub trait TransactionSigner {
     fn sign_txn(&self, raw_txn: RawTransaction) -> Result<SignedTransaction>;
 }
 
-pub trait ChannelPayloadSigner {
-    fn sign_script_payload(&self, channel_payload: &ChannelScriptBody) -> Result<Ed25519Signature> {
-        self.sign_bytes(lcs::to_bytes(channel_payload)?)
-    }
-
-    fn sign_write_set_payload(
-        &self,
-        channel_payload: &ChannelWriteSetBody,
-    ) -> Result<Ed25519Signature> {
-        self.sign_bytes(lcs::to_bytes(channel_payload)?)
-    }
-
-    fn sign_bytes(&self, bytes: Vec<u8>) -> Result<Ed25519Signature>;
+pub trait ChannelTransactionSigner {
+    /// sign method should return (participant index, participant signature)
+    fn sign(&self, body: &ChannelTransactionPayloadBodyV2) -> (usize, Ed25519Signature);
 }
 
 /// Craft a transaction request.
@@ -136,15 +124,5 @@ impl TransactionSigner for KeyPair<Ed25519PrivateKey, Ed25519PublicKey> {
             self.public_key.clone(),
             signature,
         ))
-    }
-}
-
-impl ChannelPayloadSigner for KeyPair<Ed25519PrivateKey, Ed25519PublicKey> {
-    fn sign_bytes(&self, bytes: Vec<u8>) -> Result<Ed25519Signature> {
-        //TODO(jole) use a special hasher.
-        let mut hasher = TestOnlyHasher::default();
-        hasher.write(bytes.as_slice());
-        let hash = hasher.finish();
-        Ok(self.private_key.sign_message(&hash))
     }
 }
